Modified Scripts and Prefabs
I changed the "mesh changing" script, it changes the texture and the
shader of the item as well. Also, I changed the "Ball Changer" prefab,
it now has a hovering hamsterball which indicates the texture/shader
that the ball will change to upon trigger enter. I modified the alien
patrol prefab. I also created a portal prefab with texture that is used
to end the level. Included is some extra textures such as rubber. Whew,
that was a lot...