Made a switch
Switch has a light attached, upon triggering switch the button descends
and the light changes to blue. The gameObject that it deactivates is an
array. You can add as many gameObjects that you want to deactivate to
the script in unity. Also, to deactivate a laser/electric wall, attach
the wall part of the prefab. I messed with the electric bounce script,
but I hate it... soo.... yeah...