Made this game my BITCH! Wewt!
Okay! So this commit actually has a lot of shit in it.
1. I fixed laser walls or electric walls or whatever. The script is in
laser walls but the fix was way easier than we thought. Anyway, they
dont push the hamster out anymore.
2. I created a better portal prefab. We decided to go with collect all
of the collectibles before you can move on.  I feel this encourages the
puzzling aspect. When a collectible is picked up, 1 of 3 lights light up
on the portal. Once all collectibles have been collected the player can
continue to the next level.
3. I changed the collider with the switch so that the player doesn't
miss all the time.
4. I changed the timer on the ball changer from 0.5 to 1, this lets the
player linger a little and not deal with unwanted switching.
Thats about it!!! The game is looking good!